iOS - fixing sending the stream
As per #26, extending native MediaStream doesn't work.

With this fix, I'm able to stream from my iPad.

For whatever reason, I cannot receive the stream. I don't know if that's my app, or there is still something that is missing.

There are no errors in the console that I can see in my app. Could someone test it, please?