Resolve `join` promise after API channel ready
Previously, the (implicit) promise returned by `join` would resolve
after establishing the transports, but before the API data channel was
attached.  This meant there was a small window of time where the API
data channel was not ready, and trying to use the subscriber API during
that time would fail silently.

This change returns a promise that does not resolve until after the API
channel is established.